README.md
Jamshid Khaksaar was born in Kunduz in 1989, due to the wars we left for Iran and come back in 2001 to Afghanistan. 
From my childhood time, I was crushed by Technology and loved to work with computers and machines, meanwhile, I was a super fan of Super Mario. 
In 2007 when I graduated from School I start thinking to become a programmer, I am trying and I will keep it up till I become a full-stack developer. 
I am a fan of open-source projects and Linux. 
Yes, I am: 
🎮 A Gamer, 👨‍🏫 A Self Learner, and Becoming and Pro Programmer Soon 💻